fire range adjusted, range shown when focused

+++ app/src/main/java/com/thetechnoobs/typicaltowerdefence/enemys/EasyFastEnemy.java
@@ -22,7 +22,7 @@ public class EasyFastEnemy extends EnemyBase {
         this.speed = Tools.convertDpToPixel(2f);
 
         debugPaint.setColor(Color.RED);
-        debugPaint.setTextSize(Tools.convertDpToPixel(20));
+        debugPaint.setTextSize((float) Tools.convertDpToPixel(20));
     }
 
     @Override
