+++ app/src/main/java/com/thetechnoobs/typicaltowerdefence/enemys/EasySlowEnemy.java
@@ -2,15 +2,14 @@ package com.thetechnoobs.typicaltowerdefence.enemys;
 
 import android.graphics.Canvas;
 import android.graphics.Color;
-import android.graphics.Paint;
 import android.graphics.RectF;
 
 import com.thetechnoobs.typicaltowerdefence.Tools;
 
 import java.util.ArrayList;
-import java.util.Random;
 
 public class EasySlowEnemy extends EnemyBase {
+
     double distanceMoved = 0;
 
     public EasySlowEnemy(int x, int y, int[] screenSize, ArrayList<RectF> mapPath) {
@@ -22,7 +21,7 @@ public class EasySlowEnemy extends EnemyBase {
         this.speed = Tools.convertDpToPixel(1f);
 
         debugPaint.setColor(Color.RED);
-        debugPaint.setTextSize(Tools.convertDpToPixel(20));
+        debugPaint.setTextSize((float) Tools.convertDpToPixel(20));
     }
 
     @Override
@@ -43,12 +42,12 @@ public class EasySlowEnemy extends EnemyBase {
     public void draw(Canvas canvas) {
         super.draw(canvas);
 
-            debugPaint.setColor(Color.RED);
-            canvas.drawRect(getHitbox(), debugPaint);
+        debugPaint.setColor(Color.RED);
+        canvas.drawRect(getHitbox(), debugPaint);
     }
 
     public RectF getHitbox() {
-        return new RectF(getCurX(), getCurY(), getCurX() + Tools.convertDpToPixel(10), getCurY() + Tools.convertDpToPixel(10));
+        return new RectF(getCurX(), getCurY(), getCurX() + (float) Tools.convertDpToPixel(10), getCurY() + (float) Tools.convertDpToPixel(10));
     }
 }
